backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pymongo==4.10.1
pydantic==2.11.1
pydantic-settings==2.12.0
python-dotenv==1.0.1
google-genai
python-multipart
